+++ src/main/java/com/fooeating/db/PublicDAO.java
@@ -11,7 +11,7 @@ import javax.naming.Context;
 import javax.naming.InitialContext;
 import javax.sql.DataSource;
 
-public class AdminDAO {
+public class PublicDAO {
 	private Connection con = null;
 	private PreparedStatement pstmt = null;
 	private ResultSet rs = null;
@@ -68,5 +68,35 @@ public class AdminDAO {
 		
 		return userList;
 	}
+	
+	// 회원가입 - MemberJoin()
+		public void MemberJoin(UserDTO dto) {
+			try {
+				// 1.2. 디비연결
+				con = getCon();
+				// 3. sql작성&pstmt 객체
+				sql = "insert into user(user_id,pw,name,email,phone,regdate) "
+						+ " values(?,?,?,?,?,?)";
+				pstmt = con.prepareStatement(sql);
+				pstmt.setString(1, dto.getId());
+				pstmt.setString(2, dto.getPw());
+				pstmt.setString(3, dto.getName());
+				pstmt.setString(4, dto.getEmail());
+				pstmt.setString(5, dto.getPhone());
+				pstmt.setTimestamp(6, dto.getRegdate());
+				
+				// 4. sql 실행
+				pstmt.executeUpdate();
+				System.out.println(" DAO : 회원가입 성공!");
+				
+				
+			} catch (Exception e) {
+				e.printStackTrace();
+			}finally {
+				closeDB();
+			}
+			
+			
+		}// 회원가입 - MemberJoin()
 	
 }
